Updating clan info

+++ src/Models/Clan/PacketClanInfoReadResult.cs
@@ -42,8 +42,8 @@ namespace Kamael.Packets.Clan
             ElderCount = packet.ReadUInt16();
             RequireApproval = packet.ReadByte();
             LevelRequirement = packet.ReadUInt16();
-            CastleOwned = packet.ReadUInt32();
-            FortressOwned = Convert.ToString(packet.ReadInt64());
+            CastleOwned = packet.ReadUInt16();
+            FortressOwned = Convert.ToString(packet.ReadInt32());
             RedStarStone = packet.ReadUInt32();
             packet.Skip(100); // Junk Data
             PacketID2 = packet.ReadUInt32();
